rename: 게시글 repository를 Accompany 게시글 repository로 변경(#12)
