+++ src/main/java/com/example/gabojago_server/repository/article/accompany/AccompanyArticleRepository.java
@@ -1,4 +1,4 @@
-package com.example.gabojago_server.repository.article;
+package com.example.gabojago_server.repository.article.accompany;
 
 import com.example.gabojago_server.model.article.AccompanyArticle;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ArticleRepository extends JpaRepository<AccompanyArticle, Long>, ArticleRepositoryCustom {
+public interface AccompanyArticleRepository extends JpaRepository<AccompanyArticle, Long>, AccompanyArticleRepositoryCustom {
 
     @Query("select count(a) > 0 from Article a where a.writer.id = :writer and  a.id = :article")
     boolean existArticleByWriter(@Param("writer") Long writerId, @Param("article") Long articleId);
