+++ src/main/java/com/example/gabojago_server/repository/article/accompany/AccompanyArticleRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.example.gabojago_server.repository.article;
+package com.example.gabojago_server.repository.article.accompany;
 
 import com.example.gabojago_server.dto.response.article.PageAccompanyResponseDto;
 import org.springframework.data.domain.Page;
@@ -6,6 +6,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ArticleRepositoryCustom {
+public interface AccompanyArticleRepositoryCustom {
     Page<PageAccompanyResponseDto> searchAll(Pageable pageable);
 }
