+++ src/test/java/com/example/gabojago_server/service/comment/CommentServiceTest.java
@@ -5,7 +5,7 @@ import com.example.gabojago_server.dto.response.comment.CommentResponseDto;
 import com.example.gabojago_server.model.article.AccompanyArticle;
 import com.example.gabojago_server.model.article.Article;
 import com.example.gabojago_server.model.member.Member;
-import com.example.gabojago_server.repository.article.ArticleRepository;
+import com.example.gabojago_server.repository.article.accompany.AccompanyArticleRepository;
 import com.example.gabojago_server.repository.articlecomment.ArticleCommentRepository;
 import com.example.gabojago_server.repository.member.MemberRepository;
 import org.junit.jupiter.api.BeforeEach;
@@ -28,7 +28,7 @@ class CommentServiceTest {
     CommentService commentService;
 
     @Autowired
-    ArticleRepository articleRepository;
+    AccompanyArticleRepository articleRepository;
 
     @Autowired
     MemberRepository memberRepository;
